+++ Fronter/Source/Configuration/Configuration.cpp
@@ -4,6 +4,7 @@
 #include <filesystem>
 #include <fstream>
 namespace fs = std::filesystem;
+#include "../Utils/CommonFunctions.h"
 
 Configuration::Configuration()
 {
@@ -156,101 +157,6 @@ bool Configuration::exportConfiguration() const
 
 bool Configuration::copyMod() const
 {
-	if (!fs::exists(converterFolder))
-	{
-		Log(LogLevel::Error) << "Copy failed - where is the converter?";
-		return false;
-	}
-	if (!fs::exists(converterFolder + "/output"))
-	{
-		Log(LogLevel::Error) << "Copy failed - where is the converter's output folder?";
-		return false;
-	}
-
-	const auto& folderItr = requiredFolders.find("targetGameModPath");
-	if (folderItr == requiredFolders.end())
-	{
-		Log(LogLevel::Error) << "Copy failed - Target Folder isn't loaded!";
-		return false;
-	}
-	const auto& destinationFolder = folderItr->second->getValue();
-	if (!fs::exists(fs::u8path(destinationFolder)))
-	{
-		Log(LogLevel::Error) << "Copy failed - Target Folder does not exist!";
-		return false;
-	}
-	std::string targetName;
-	for (const auto& option: options)
-	{
-		if (option->getName() == "output_name" && !option->getValue().empty())
-		{
-			targetName = option->getValue();
-		}
-	}
-	if (targetName.empty())
-	{
-		const auto& fileItr = requiredFiles.find("SaveGame");
-		if (fileItr == requiredFiles.end())
-		{
-			Log(LogLevel::Error) << "Copy failed - SaveGame is does not exist!";
-			return false;
-		}
-		auto saveGamePath = fileItr->second->getValue();
-		if (saveGamePath.empty())
-		{
-			Log(LogLevel::Error) << "Copy Failed - save game path is empty, did we even convert anything?";
-			return false;
-		}
-		if (!fs::exists(fs::u8path(saveGamePath)))
-		{
-			Log(LogLevel::Error) << "Copy Failed - save game does not exist, did we even convert anything?";
-			return false;
-		}
-		if (fs::is_directory(fs::u8path(saveGamePath)))
-		{
-			Log(LogLevel::Error) << "Copy Failed - Save game is a directory...";
-			return false;
-		}
-		saveGamePath = trimPath(saveGamePath);
-		saveGamePath = normalizeStringPath(saveGamePath);
-		const auto pos = saveGamePath.find_last_of('.');
-		if (pos != std::string::npos)
-			saveGamePath = saveGamePath.substr(0, pos);
-		targetName = saveGamePath;
-	}
-	if (!fs::exists(fs::u8path(converterFolder + "/output/" + targetName + ".mod")))
-	{
-		Log(LogLevel::Error) << "Copy Failed - Cound not find mod: " << converterFolder + "/output/" + targetName + ".mod";
-		return false;
-	}
-	if (!fs::exists(fs::u8path(converterFolder + "/output/" + targetName)))
-	{
-		Log(LogLevel::Error) << "Copy Failed - Cound not find mod folder: " << converterFolder + "/output/" + targetName;
-		return false;
-	}
-	if (!fs::is_directory(fs::u8path(converterFolder + "/output/" + targetName)))
-	{
-		Log(LogLevel::Error) << "Copy Failed - Mod folder is not a directory: " << converterFolder + "/output/" + targetName;
-		return false;
-	}
-	if (fs::exists(fs::u8path(destinationFolder + "/" + targetName + ".mod")))
-		fs::remove(fs::u8path(destinationFolder + "/" + targetName + ".mod"));
-	if (fs::exists(fs::u8path(destinationFolder + "/" + targetName)))
-		fs::remove_all(fs::u8path(destinationFolder + "/" + targetName));
-	try
-	{
-		fs::copy(fs::u8path(converterFolder + "/output/" + targetName + ".mod"), fs::u8path(destinationFolder + "/" + targetName + ".mod"));
-		fs::copy(fs::u8path(converterFolder + "/output/" + targetName),
-			 fs::u8path(destinationFolder + "/" + targetName),
-			 std::filesystem::copy_options::recursive);
-	}
-	catch (std::filesystem::filesystem_error& theError)
-	{
-		Log(LogLevel::Error) << theError.what();
-		return false;
-	}
-	Log(LogLevel::Info) << "Mod successfully copied to: " << destinationFolder + "/" + targetName;
-	return true;
 }
 
 std::string Configuration::getSecondTailSource() const
@@ -258,63 +164,3 @@ std::string Configuration::getSecondTailSource() const
 	return converterFolder + "/log.txt";
 }
 
-LogMessage Configuration::Configuration::sliceMessage(const std::string& message)
-{
-	LogMessage logMessage;
-	const auto posOpen = message.find_first_of('[');
-	if (!(posOpen >= 20 && posOpen <= 23))
-	{
-		// This is not a standard message. File as info ad verbatim.
-		logMessage.message = message;
-		return logMessage;
-	}
-	const auto posClose = message.find_first_of(']');
-	if (posClose == std::string::npos)
-	{
-		// something's very wrong with this message.
-		logMessage.message = message;
-		return logMessage;
-	}
-	const auto logLevel = message.substr(posOpen + 1, posClose - posOpen - 1);
-	if (logLevel == "INFO")
-		logMessage.logLevel = LogLevel::Info;
-	else if (logLevel == "DEBUG")
-		logMessage.logLevel = LogLevel::Debug;
-	else if (logLevel == "WARNING")
-		logMessage.logLevel = LogLevel::Warning;
-	else if (logLevel == "ERROR")
-		logMessage.logLevel = LogLevel::Error;
-	logMessage.timestamp = message.substr(0, 19);
-	logMessage.message = message.substr(posClose + 2, message.length());
-	return logMessage;
-}
-
-std::string Configuration::normalizeStringPath(const std::string& stringPath)
-{
-	std::string toReturn = Utils::normalizeUTF8Path(stringPath);
-	toReturn = replaceCharacter(toReturn, '-');
-	toReturn = replaceCharacter(toReturn, ' ');
-
-	return toReturn;
-}
-
-std::string Configuration::replaceCharacter(std::string fileName, char character)
-{
-	auto position = fileName.find_first_of(character);
-	while (position != std::string::npos)
-	{
-		fileName.replace(position, 1, "_");
-		position = fileName.find_first_of(character);
-	}
-
-	return fileName;
-}
-
-std::string Configuration::trimPath(const std::string& fileName)
-{
-	auto lastSlash = fileName.find_last_of('\\');
-	auto trimmedFileName = fileName.substr(static_cast<int>(lastSlash) + 1, fileName.length());
-	lastSlash = trimmedFileName.find_last_of('/');
-	trimmedFileName = trimmedFileName.substr(static_cast<int>(lastSlash) + 1, trimmedFileName.length());
-	return trimmedFileName;
-}
