+++ Fronter/Source/Configuration/Configuration.h
@@ -6,21 +6,6 @@
 #include "RequiredFolder.h"
 #include "Parser.h"
 
-enum class MessageSource
-{
-	UNINITIALIZED = 0,
-	UI = 1,
-	CONVERTER = 2
-};
-
-typedef struct
-{
-	std::string timestamp;
-	LogLevel logLevel = LogLevel::Info;
-	MessageSource source = MessageSource::UNINITIALIZED;
-	std::string message;
-} LogMessage;
-
 class Configuration: commonItems::parser
 {
   public:
@@ -33,16 +18,12 @@ class Configuration: commonItems::parser
 	[[nodiscard]] const auto& getDisplayName() const { return displayName; }
 	[[nodiscard]] const auto& getRequiredFiles() const { return requiredFiles; }
 	[[nodiscard]] const auto& getRequiredFolders() const { return requiredFolders; }
+	[[nodiscard]] const auto& getConverterFolder() const { return converterFolder; }
 	[[nodiscard]] const auto& getOptions() const { return options; }
 
 	[[nodiscard]] std::string getSecondTailSource() const;
 	[[nodiscard]] bool exportConfiguration() const;
 	[[nodiscard]] bool copyMod() const;
-
-	static LogMessage sliceMessage(const std::string& message);
-	static std::string normalizeStringPath(const std::string& stringPath);
-	static std::string replaceCharacter(std::string fileName, char character);
-	static std::string trimPath(const std::string& fileName);
 
   private:
 	void registerKeys();
